+++ simpletransformers/classification/classification_utils.py
@@ -17,18 +17,17 @@
 
 from __future__ import absolute_import, division, print_function
 
+import csv
+import json
+import logging
 import os
 import sys
-import csv
-import logging
-
 from io import open
 from multiprocessing import Pool, cpu_count
 
-from tqdm.auto import tqdm
 from scipy.stats import pearsonr, spearmanr
-from sklearn.metrics import matthews_corrcoef, f1_score
-
+from tqdm.auto import tqdm
+import shutil
 
 logger = logging.getLogger(__name__)
 csv.field_size_limit(2147483647)
@@ -68,16 +67,30 @@ class InputFeatures(object):
 
 
 def convert_example_to_feature(
-        example_row,
-        pad_token=0,
-        sequence_a_segment_id=0,
-        sequence_b_segment_id=1,
-        cls_token_segment_id=1,
-        pad_token_segment_id=0,
-        mask_padding_with_zero=True,
-        sep_token_extra=False
-    ):
-    example, max_seq_length, tokenizer, output_mode, cls_token_at_end, cls_token, sep_token, cls_token_segment_id, pad_on_left, pad_token_segment_id, sep_token_extra, multi_label, stride = example_row
+    example_row,
+    pad_token=0,
+    sequence_a_segment_id=0,
+    sequence_b_segment_id=1,
+    cls_token_segment_id=1,
+    pad_token_segment_id=0,
+    mask_padding_with_zero=True,
+    sep_token_extra=False,
+):
+    (
+        example,
+        max_seq_length,
+        tokenizer,
+        output_mode,
+        cls_token_at_end,
+        cls_token,
+        sep_token,
+        cls_token_segment_id,
+        pad_on_left,
+        pad_token_segment_id,
+        sep_token_extra,
+        multi_label,
+        stride,
+    ) = example_row
 
     tokens_a = tokenizer.tokenize(example.text_a)
 
@@ -93,7 +106,7 @@ def convert_example_to_feature(
         # Account for [CLS] and [SEP] with "- 2" and with "- 3" for RoBERTa.
         special_tokens_count = 3 if sep_token_extra else 2
         if len(tokens_a) > max_seq_length - special_tokens_count:
-            tokens_a = tokens_a[:(max_seq_length - special_tokens_count)]
+            tokens_a = tokens_a[: (max_seq_length - special_tokens_count)]
 
     # The convention in BERT is:
     # (a) For sequence pairs:
@@ -137,11 +150,15 @@ def convert_example_to_feature(
     padding_length = max_seq_length - len(input_ids)
     if pad_on_left:
         input_ids = ([pad_token] * padding_length) + input_ids
-        input_mask = ([0 if mask_padding_with_zero else 1] * padding_length) + input_mask
+        input_mask = (
+            [0 if mask_padding_with_zero else 1] * padding_length
+        ) + input_mask
         segment_ids = ([pad_token_segment_id] * padding_length) + segment_ids
     else:
         input_ids = input_ids + ([pad_token] * padding_length)
-        input_mask = input_mask + ([0 if mask_padding_with_zero else 1] * padding_length)
+        input_mask = input_mask + (
+            [0 if mask_padding_with_zero else 1] * padding_length
+        )
         segment_ids = segment_ids + ([pad_token_segment_id] * padding_length)
 
     assert len(input_ids) == max_seq_length
@@ -159,21 +176,35 @@ def convert_example_to_feature(
         input_ids=input_ids,
         input_mask=input_mask,
         segment_ids=segment_ids,
-        label_id=example.label
+        label_id=example.label,
     )
 
 
 def convert_example_to_feature_sliding_window(
-        example_row,
-        pad_token=0,
-        sequence_a_segment_id=0,
-        sequence_b_segment_id=1,
-        cls_token_segment_id=1,
-        pad_token_segment_id=0,
-        mask_padding_with_zero=True,
-        sep_token_extra=False,
-    ):
-    example, max_seq_length, tokenizer, output_mode, cls_token_at_end, cls_token, sep_token, cls_token_segment_id, pad_on_left, pad_token_segment_id, sep_token_extra, multi_label, stride = example_row
+    example_row,
+    pad_token=0,
+    sequence_a_segment_id=0,
+    sequence_b_segment_id=1,
+    cls_token_segment_id=1,
+    pad_token_segment_id=0,
+    mask_padding_with_zero=True,
+    sep_token_extra=False,
+):
+    (
+        example,
+        max_seq_length,
+        tokenizer,
+        output_mode,
+        cls_token_at_end,
+        cls_token,
+        sep_token,
+        cls_token_segment_id,
+        pad_on_left,
+        pad_token_segment_id,
+        sep_token_extra,
+        multi_label,
+        stride,
+    ) = example_row
 
     if stride < 1:
         stride = int(max_seq_length * stride)
@@ -185,13 +216,16 @@ def convert_example_to_feature_sliding_window(
 
     special_tokens_count = 3 if sep_token_extra else 2
     if len(tokens_a) > bucket_size:
-        token_sets = [tokens_a[i:i + bucket_size] for i in range(0, len(tokens_a), stride)]
+        token_sets = [
+            tokens_a[i : i + bucket_size] for i in range(0, len(tokens_a), stride)
+        ]
     else:
         token_sets.append(tokens_a)
 
-
     if example.text_b:
-        raise ValueError("Sequence pair tasks not implemented for sliding window tokenization.")
+        raise ValueError(
+            "Sequence pair tasks not implemented for sliding window tokenization."
+        )
 
     # The convention in BERT is:
     # (a) For sequence pairs:
@@ -234,11 +268,15 @@ def convert_example_to_feature_sliding_window(
         padding_length = max_seq_length - len(input_ids)
         if pad_on_left:
             input_ids = ([pad_token] * padding_length) + input_ids
-            input_mask = ([0 if mask_padding_with_zero else 1] * padding_length) + input_mask
+            input_mask = (
+                [0 if mask_padding_with_zero else 1] * padding_length
+            ) + input_mask
             segment_ids = ([pad_token_segment_id] * padding_length) + segment_ids
         else:
             input_ids = input_ids + ([pad_token] * padding_length)
-            input_mask = input_mask + ([0 if mask_padding_with_zero else 1] * padding_length)
+            input_mask = input_mask + (
+                [0 if mask_padding_with_zero else 1] * padding_length
+            )
             segment_ids = segment_ids + ([pad_token_segment_id] * padding_length)
 
         assert len(input_ids) == max_seq_length
@@ -257,7 +295,7 @@ def convert_example_to_feature_sliding_window(
                 input_ids=input_ids,
                 input_mask=input_mask,
                 segment_ids=segment_ids,
-                label_id=example.label
+                label_id=example.label,
             )
         )
 
@@ -265,29 +303,29 @@ def convert_example_to_feature_sliding_window(
 
 
 def convert_examples_to_features(
-        examples,
-        max_seq_length,
-        tokenizer,
-        output_mode,
-        cls_token_at_end=False,
-        sep_token_extra=False,
-        pad_on_left=False,
-        cls_token="[CLS]",
-        sep_token="[SEP]",
-        pad_token=0,
-        sequence_a_segment_id=0,
-        sequence_b_segment_id=1,
-        cls_token_segment_id=1,
-        pad_token_segment_id=0,
-        mask_padding_with_zero=True,
-        process_count=cpu_count() - 2,
-        multi_label=False,
-        silent=False,
-        use_multiprocessing=True,
-        sliding_window=False,
-        flatten=False,
-        stride=None
-    ):
+    examples,
+    max_seq_length,
+    tokenizer,
+    output_mode,
+    cls_token_at_end=False,
+    sep_token_extra=False,
+    pad_on_left=False,
+    cls_token="[CLS]",
+    sep_token="[SEP]",
+    pad_token=0,
+    sequence_a_segment_id=0,
+    sequence_b_segment_id=1,
+    cls_token_segment_id=1,
+    pad_token_segment_id=0,
+    mask_padding_with_zero=True,
+    process_count=cpu_count() - 2,
+    multi_label=False,
+    silent=False,
+    use_multiprocessing=True,
+    sliding_window=False,
+    flatten=False,
+    stride=None,
+):
     """ Loads a data file into a list of `InputBatch`s
         `cls_token_at_end` define the location of the CLS token:
             - False (Default, BERT/XLM pattern): [CLS] + A + [SEP] + B + [SEP]
@@ -295,28 +333,71 @@ def convert_examples_to_features(
         `cls_token_segment_id` define the segment id associated to the CLS token (0 for BERT, 2 for XLNet)
     """
 
-    examples = [(example, max_seq_length, tokenizer, output_mode, cls_token_at_end, cls_token, sep_token, cls_token_segment_id, pad_on_left, pad_token_segment_id, sep_token_extra, multi_label, stride) for example in examples]
+    examples = [
+        (
+            example,
+            max_seq_length,
+            tokenizer,
+            output_mode,
+            cls_token_at_end,
+            cls_token,
+            sep_token,
+            cls_token_segment_id,
+            pad_on_left,
+            pad_token_segment_id,
+            sep_token_extra,
+            multi_label,
+            stride,
+        )
+        for example in examples
+    ]
 
     if use_multiprocessing:
         if sliding_window:
-            print('sliding_window enabled')
+            print("sliding_window enabled")
             with Pool(process_count) as p:
-                features = list(tqdm(p.imap(convert_example_to_feature_sliding_window, examples, chunksize=500), total=len(examples), disable=silent))
+                features = list(
+                    tqdm(
+                        p.imap(
+                            convert_example_to_feature_sliding_window,
+                            examples,
+                            chunksize=500,
+                        ),
+                        total=len(examples),
+                        disable=silent,
+                    )
+                )
             if flatten:
-                features = [feature for feature_set in features for feature in feature_set]
-            print(f'{len(features)} features created from {len(examples)} samples.')
+                features = [
+                    feature for feature_set in features for feature in feature_set
+                ]
+            print(f"{len(features)} features created from {len(examples)} samples.")
         else:
             with Pool(process_count) as p:
-                features = list(tqdm(p.imap(convert_example_to_feature, examples, chunksize=500), total=len(examples), disable=silent))
+                features = list(
+                    tqdm(
+                        p.imap(convert_example_to_feature, examples, chunksize=500),
+                        total=len(examples),
+                        disable=silent,
+                    )
+                )
     else:
         if sliding_window:
-            print('sliding_window enabled')
-            features = [convert_example_to_feature_sliding_window(example) for example in tqdm(examples, disable=silent)]
+            print("sliding_window enabled")
+            features = [
+                convert_example_to_feature_sliding_window(example)
+                for example in tqdm(examples, disable=silent)
+            ]
             if flatten:
-                features = [feature for feature_set in features for feature in feature_set]
-            print(f'{len(features)} features created from {len(examples)} samples.')
+                features = [
+                    feature for feature_set in features for feature in feature_set
+                ]
+            print(f"{len(features)} features created from {len(examples)} samples.")
         else:
-            features = [convert_example_to_feature(example) for example in tqdm(examples, disable=silent)]
+            features = [
+                convert_example_to_feature(example)
+                for example in tqdm(examples, disable=silent)
+            ]
 
     return features
 
@@ -337,3 +418,40 @@ def _truncate_seq_pair(tokens_a, tokens_b, max_length):
             tokens_a.pop()
         else:
             tokens_b.pop()
+
+
+def read_json(file_path):
+    with open(file_path) as json_file:
+        data = json.load(json_file)
+    return data
+
+
+def write_json(file_path, data, encoding="utf8"):
+    with open(file_path, "w") as outfile:
+        json.dump(data, outfile, ensure_ascii=False, indent=4)
+
+
+def update_results_file(results, results_path, output_dir_current):
+    if os.path.exists(results_path):
+        dicc = read_json(results_path)
+    else:
+        dicc = {}
+    dicc[os.path.basename(output_dir_current)] = results
+    write_json(results_path, dicc)
+
+
+def delete_worst_models(args, results_path):  
+    results_list = []
+    results_dict = read_json(results_path)
+    # Save to list 
+    for result in results_dict.keys():
+        metric_val = results_dict[result][args["metric_criteria"]]
+        results_list.append([result,metric_val])
+    print(results_list)
+    # Sort results
+    results_list.sort(key=lambda x: x[1],reverse=True)
+    # Delete worst epochs
+    for item in results_list[int(args["save_n_best_epochs"]):]:
+        epoch_path= os.path.join(args["output_dir"],item[0])
+        if os.path.exists(epoch_path):
+            shutil.rmtree(epoch_path)